Move functions from histogram.py into hi_l1a.py
